Prevent HTML parsing in appendText (#41)
* Prevent HTML parsing in appendText

Currently, if you pass an HTML-looking string to `appendText`, the
text will be parsed as HTML. This was a problem for TryPureScript,
as the errors were rendered using this function. Thus, if you put
HTML in a type error, it would be executed, which led to this
fantastic moment:

https://gist.github.com/i-am-tom/cc552efe2f82c7c42d63c6b0ec69e614

This commit uses `document.createTextNode` to prevent this from
happening, which essentially does a pauper's HTML escape, as per
this StackOverflow answer: https://stackoverflow.com/a/944456

* Update as per @paf31's suggestions

The original function has been preserved and renamed to
`unsafeAppendHTML`, with the new "safe" version being `appendText`.